requirements.txt
google-api-python-client>=2.0.0
google-auth>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
pytrends>=4.9.0
requests>=2.28.0
